requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
paho-mqtt
websockets
aiohttp
python-multipart
jinja2
